requirements.txt
rumps>=0.4.0
PyYAML>=6.0
pyobjc-core>=9.0
pyobjc-framework-Cocoa>=9.0
setuptools==67.6.1
py2app>=0.28.4
